race GUI refactoring nearly completed with InfoPanel implementation
